scraper/requirements.txt
beautifulsoup4
requests
fastapi
uvicorn
python-dotenv
pyhamcrest